Move the Cu about to fusion folder
