+++ src/stories/components/NavigationCard/CardExpenses.tsx
@@ -3,11 +3,11 @@ import { Typography, useMediaQuery } from '@mui/material';
 import { siteRoutes } from '@ses/config/routes';
 import { ResourceType } from '@ses/core/models/interfaces/types';
 import React from 'react';
+import { DividerStyle } from '@/views/CUAbout/CuAboutContainer';
 import lightTheme from '../../../../styles/theme/themes';
 import { useThemeContext } from '../../../core/context/ThemeContext';
 import { ButtonType } from '../../../core/enums/buttonTypeEnum';
 import { MAKER_BURN_LINK } from '../../../core/utils/const';
-import { DividerStyle } from '../../containers/CUAbout/CuAboutContainer';
 import InlineUser from '../../containers/TransparencyReport/components/InlineUser/InlineUser';
 import { CustomLink } from '../CustomLink/CustomLink';
 import { LinkButton } from '../LinkButton/LinkButton';
