+++ src/views/CUAbout/CuAboutContainer.stories.tsx
@@ -4,12 +4,12 @@ import { withoutSBPadding } from '@ses/core/utils/storybook/decorators';
 import { createThemeModeVariants } from '@ses/core/utils/storybook/factories';
 import { SESCoreUnitMocked } from '@ses/core/utils/storybook/mocks/coreUnitsMocks';
 import { featureFlags } from 'feature-flags/feature-flags';
-import AppLayout from '../AppLayout/AppLayout';
+import AppLayout from '@/stories/containers/AppLayout/AppLayout';
 import CuAboutContainer from './CuAboutContainer';
 import type { Meta } from '@storybook/react';
 
 const meta: Meta<typeof CuAboutContainer> = {
-  title: 'Pages/CU About',
+  title: 'Fusion/Pages/CU_About',
   component: CuAboutContainer,
   decorators: [withoutSBPadding],
