+++ src/views/CUAbout/useCuAbout.ts
@@ -1,9 +1,9 @@
 import sortBy from 'lodash/sortBy';
 import { useMemo, useCallback, useRef } from 'react';
+import { getRelateMipObjectFromCoreUnit } from '@/core/businessLogic/coreUnitAbout';
 import { TeamStatus } from '@/core/models/interfaces/types';
-import { getRelateMipObjectFromCoreUnit } from '../../../core/businessLogic/coreUnitAbout';
-import { getArrayParam, getStringParam } from '../../../core/utils/filters';
-import { buildQueryString } from '../../../core/utils/urls';
+import { getArrayParam, getStringParam } from '@/core/utils/filters';
+import { buildQueryString } from '@/core/utils/urls';
 import type { CoreUnit } from '@ses/core/models/interfaces/coreUnit';
 import type { CuMip } from '@ses/core/models/interfaces/cuMip';
 import type { NextRouter } from 'next/router';
